Updated test suite, fixed some bugs

+++ features/step_definitions/laze_steps.rb
@@ -5,11 +5,11 @@ end
 
 After do
   Dir.chdir(TEST_DIR)
-  FileUtils.rm_rf(TEST_DIR)
+  FileUtils.rm_rf(TEST_DIR) if File.directory?(TEST_DIR)
 end
 
-Given /^I have an '(.+?)' file that contains '(.+?)'$/ do |file, text|
-  FileUtils.mkdir_p(File.dirname(file)) # unless File.directory?('input')
+Given /^I have an '(.+?)' file that contains ('|")(.*)\2$/ do |file, quote, text|
+  FileUtils.mkdir_p(File.dirname(file))
   File.open(file, 'w') do |f|
     f.write(text)
   end
@@ -19,7 +19,7 @@ Given /^I have a (.+?) directory$/ do |dir|
   FileUtils.mkdir(dir)
 end
 
-Given /^I have a '(.+?)' layout that contains '(.*)'$/ do |layout, text|
+Given /^I have a '(.+?)' layout that contains ('|")(.*)\2$/ do |layout, quote, text|
   File.open(File.join('layouts', layout + '.html'), 'w') do |f|
     f.write(text)
   end
@@ -41,6 +41,13 @@ Given /^I have a '(.+?)' file that contains '(.*)'$/ do |file, text|
   end
 end
 
+Given /^I have an? '(.+?)' file that contains$/ do |file, text|
+  FileUtils.mkdir_p(File.dirname(file))
+  File.open(file, 'w') do |f|
+    f.write(text)
+  end
+end
+
 Given /^I have an '(.*?)' page(?: with (.*?) '(.*?)')?(?: and with (.*?) '(.*?)')? that contains '(.*)'$/ do |file, key, value, other_key, other_value, text|
   FileUtils.mkdir('input') unless File.directory?('input')
   File.open(file, 'w') do |f|
@@ -55,6 +62,10 @@ end
 
 When /^I run laze$/ do
   run_laze
+end
+
+When /^I run laze with minification$/ do
+  run_laze(:minify => true)
 end
 
 Then /^the (.+?) directory should exist$/ do |dir|
